Medical staff api utils

+++ backend/src/main/java/com/csse/healthSphere/dto/MedicalStaffRequest.java
@@ -6,6 +6,6 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = true)
 @Data
 public class MedicalStaffRequest extends PersonRequest {
-    int departmentId;
-    int wardId;
+    Long departmentId;
+    Long wardId;
 }
